+++ packages/engine/Source/Widget/CesiumWidget.js
@@ -172,7 +172,6 @@ function configureCameraFrustum(widget) {
  * @param {boolean} [options.requestRenderMode=false] If true, rendering a frame will only occur when needed as determined by changes within the scene. Enabling improves performance of the application, but requires using {@link Scene#requestRender} to render a new frame explicitly in this mode. This will be necessary in many cases after making changes to the scene in other parts of the API. See {@link https://cesium.com/blog/2018/01/24/cesium-scene-rendering-performance/|Improving Performance with Explicit Rendering}.
  * @param {number} [options.maximumRenderTimeChange=0.0] If requestRenderMode is true, this value defines the maximum change in simulation time allowed before a render is requested. See {@link https://cesium.com/blog/2018/01/24/cesium-scene-rendering-performance/|Improving Performance with Explicit Rendering}.
  * @param {number} [options.msaaSamples=4] If provided, this value controls the rate of multisample antialiasing. Typical multisampling rates are 2, 4, and sometimes 8 samples per pixel. Higher sampling rates of MSAA may impact performance in exchange for improved visual quality. This value only applies to WebGL2 contexts that support multisample render targets. Set to 1 to disable MSAA.
- * @param {string} [options.cspStyleNonce] A nonce to be used for all inline styles created by the widget, for Content Security Policy compliance.
  *
  * @exception {DeveloperError} Element with id "container" does not exist in the document.
  *
@@ -328,7 +327,6 @@ function CesiumWidget(container, options) {
       maximumRenderTimeChange: options.maximumRenderTimeChange,
       depthPlaneEllipsoidOffset: options.depthPlaneEllipsoidOffset,
       msaaSamples: options.msaaSamples,
-      cspStyleNonce: options.cspStyleNonce,
     });
     this._scene = scene;