+++ packages/widgets/Source/Animation/Animation.js
@@ -1,3 +1,4 @@
+/* global CESIUM_CSP_STYLE_NONCE */
 import {
   Color,
   defined,
@@ -406,7 +407,6 @@ SvgButton.prototype.setTooltip = function (tooltip) {
  *
  * @param {Element|string} container The DOM element or ID that will contain the widget.
  * @param {AnimationViewModel} viewModel The view model used by this widget.
- * @param {string} [cspStyleNonce] A nonce to be added to the style tag for CSP compliance
  *
  * @exception {DeveloperError} Element with id "container" does not exist in the document.
  *
@@ -429,7 +429,7 @@ SvgButton.prototype.setTooltip = function (tooltip) {
  * @see AnimationViewModel
  * @see Clock
  */
-function Animation(container, viewModel, cspStyleNonce) {
+function Animation(container, viewModel) {
   //>>includeStart('debug', pragmas.debug);
   if (!defined(container)) {
     throw new DeveloperError("container is required.");
@@ -467,8 +467,8 @@ function Animation(container, viewModel, cspStyleNonce) {
 .cesium-animation-shuttleRingPausePointer { fill: url(#animation_shuttleRingPointerPaused); }\
 .cesium-animation-knobOuter { fill: url(#animation_knobOuter); }\
 .cesium-animation-knobInner { fill: url(#animation_knobInner); }";
-  if (defined(cspStyleNonce)) {
-    cssStyle.setAttribute("nonce", cspStyleNonce);
+  if (defined(CESIUM_CSP_STYLE_NONCE)) {
+    cssStyle.setAttribute("nonce", CESIUM_CSP_STYLE_NONCE);
   }
 
   ownerDocument.head.insertBefore(cssStyle, ownerDocument.head.childNodes[0]);