+++ CHANGES.md
@@ -10,6 +10,10 @@
 - Fixed label sizing for some fonts and characters [#9767](https://github.com/CesiumGS/cesium/issues/9767)
 - Fixed a type error when accessing the ellipsoid of a viewer [#13123](https://github.com/CesiumGS/cesium/pull/13123)
 
+#### Additions :tada:
+
+- Adds a `CESIUM_CSP_STYLE_NONCE` global for setting a nonce on style tags created by CesiumJS to comply with Content Security Policy (CSP) requirements.
+
 ## 1.137 - 2026-01-05
 
 ### @cesium/engine